+++ cordialsync/learning/episodes.py
@@ -1,13 +1,10 @@
 import itertools
-import math
 import matplotlib as mpl
-import warnings
 from typing import Tuple, Dict, List, Any, Union
 
-import cordialsync.utilities.constants as CONSTANTS
 from cordialsync.ai2thor.environment import AI2ThorEnvironmentWithGraph
 from cordialsync.ai2thor.episodes import MultiAgentAI2ThorEpisode
-from cordialsync.learning.episodes.furnmove import are_actions_coordinated, coordination_type_tensor
+from cordialsync.utilities.episodes import are_actions_coordinated, coordination_type_tensor
 from cordialsync.utilities.ai2thor import manhattan_dists_between_positions
 
 mpl.use("Agg", force=False)
@@ -488,174 +485,3 @@ class FurnLiftNApartStateEpisode(FurnLiftEpisode):
         return coordination_type_tensor(
             self.environment, self.available_actions, self.coordinated_action_checker
         )
-
-
-class FurnLiftNApartStateLoggingEpisode(FurnLiftNApartStateEpisode):
-    def __init__(
-        self,
-        env: AI2ThorEnvironmentWithGraph,
-        task_data: Dict[str, Any],
-        max_steps: int,
-        **kwargs
-    ):
-        super().__init__(env=env, task_data=task_data, max_steps=max_steps, **kwargs)
-        # Extra logging of step number at which an agent first saw the target
-        self._first_view_of_target = [self.max_steps + 1] * self.environment.num_agents
-        # Initial agent locations, to calculate the expert path lengths
-        self._initial_agent_locations = [
-            self.environment.get_agent_location(i)
-            for i in range(self.environment.num_agents)
-        ]
-        self._initial_agent_keys = [
-            self.environment.get_key(ag_loc) for ag_loc in self._initial_agent_locations
-        ]
-        goal_object = self.environment.get_object_by_id(self.object_id, agent_id=0)
-        self.object_location = dict(
-            x=round(goal_object["position"]["x"], 2),
-            z=round(goal_object["position"]["z"], 2),
-        )
-
-    @staticmethod
-    def location_l2_dist(loc1, loc2):
-        return math.sqrt((loc1["x"] - loc2["x"]) ** 2 + (loc1["z"] - loc2["z"]) ** 2)
-
-    def info(self):
-        final_agent_distance_from_target = [
-            self.location_l2_dist(
-                self.object_location, self.environment.get_agent_location(i)
-            )
-            for i in range(self.environment.num_agents)
-        ]
-
-        _dist_cache = {}
-
-        def agent_to_dist_to_target(agent_id, target_key):
-            key = (agent_id, target_key)
-            if key not in _dist_cache:
-                _dist_cache[key] = self.environment.shortest_path_length(
-                    source_state_key=self._initial_agent_keys[agent_id],
-                    goal_state_key=target_key,
-                )
-            return _dist_cache[key]
-
-        best_path_lengths = None
-        best_path_length = None
-        for good_target_group in self.target_key_groups:
-            for ordered_target_group in itertools.permutations(
-                good_target_group, len(good_target_group)
-            ):
-                path_lengths = tuple(
-                    agent_to_dist_to_target(agent_id, target_key)
-                    for agent_id, target_key in enumerate(ordered_target_group)
-                )
-                path_length = sum(path_lengths)
-                if best_path_lengths is None or best_path_length > path_length:
-                    best_path_length = best_path_length
-                    best_path_lengths = path_lengths
-
-        initial_agent_distance_from_target = [
-            self.location_l2_dist(self.object_location, ag_loc)
-            for ag_loc in self._initial_agent_locations
-        ]
-        return {
-            **super(JointNavigationEpisode, self).info(),
-            "accuracy": 1 * self._item_successfully_picked_up,
-            "pickupable_but_not_picked": self._pickupable_but_not_picked,
-            "picked_but_not_pickupable": self._picked_but_not_pickupable,
-            "picked_but_not_pickupable_distance": self._picked_but_not_pickupable_distance,
-            "picked_but_not_pickupable_visibility": self._picked_but_not_pickupable_visibility,
-            "reward": self.total_reward,
-            "first_view_of_target": self._first_view_of_target,
-            "final_distance_between_agents": self.manhattan_dists_between_agents(),
-            "best_path_to_target_length": best_path_length,
-            "path_to_target_lengths": best_path_lengths,
-            "initial_agent_distance_from_target": initial_agent_distance_from_target,
-            "final_agent_distance_from_target": final_agent_distance_from_target,
-            "object_location": self.object_location,
-            "scene_name": self.environment.scene_name,
-        }
-
-    def multi_step(self, actions_as_ints: Tuple[int, ...]) -> List[Dict[str, Any]]:
-        step_results = super(FurnLiftNApartStateLoggingEpisode, self).multi_step(
-            actions_as_ints=actions_as_ints
-        )
-
-        # Extra logging of step number at which an agent first saw the target
-        visibility = self.goal_visibility()
-        for i in range(self.environment.num_agents):
-            if visibility[i]:
-                self._first_view_of_target[i] = min(
-                    self._first_view_of_target[i], self.num_steps_taken_in_episode()
-                )
-
-        return step_results
-
-
-class FurnLiftGridStateEpisode(FurnLiftNApartStateEpisode):
-    def __init__(
-        self,
-        env: AI2ThorEnvironmentWithGraph,
-        task_data: Dict[str, Any],
-        max_steps: int,
-        **kwargs
-    ):
-        super().__init__(env=env, task_data=task_data, max_steps=max_steps, **kwargs)
-        self.object_points_set = kwargs["object_points_set"]
-
-    def states_for_agents(self) -> List[Dict[str, Any]]:
-        # TODO: Getting the current occupancy matrix like this on every iteration is,
-        # TODO: in principle, an expensive operation as it requires doing some thor actions.
-        # TODO: We could track state changes without doing this but this requires some annoying bookkeeping.
-
-        # Reachable, unreachable and agent locations are marked
-        matrix, point_to_element_map = self.environment.get_current_occupancy_matrix(
-            padding=0.5, use_initially_reachable_points_matrix=True
-        )
-        # Mark goal object locations
-        nskipped = 0
-        for point_tuple in self.object_points_set:
-            if point_tuple not in point_to_element_map:
-                nskipped += 1
-                continue
-            row, col = point_to_element_map[point_tuple]
-            matrix[row, col] = CONSTANTS.GOAL_OBJ_SYM
-
-        if nskipped == len(self.object_points_set):
-            raise RuntimeError(
-                "Skipped all object points in scene {}.".format(
-                    self.environment.scene_name
-                )
-            )
-        elif nskipped > 10:
-            warnings.warn(
-                "Skipping many object points ({}) in scene {}.".format(
-                    nskipped, self.environment.scene_name
-                )
-            )
-
-        states = []
-        for agent_id in range(self.environment.num_agents):
-            matrix_ego_per_agent = self.environment.current_matrix_frame(
-                agent_id,
-                matrix,
-                point_to_element_map,
-                int(self.environment.visibility_distance // self.environment.grid_size),
-                int(self.environment.visibility_distance // self.environment.grid_size),
-            )
-
-            last_action = (
-                None if self._last_actions is None else self._last_actions[agent_id]
-            )
-            last_action_success = (
-                None
-                if self._last_actions_success is None
-                else self._last_actions_success[agent_id]
-            )
-            states.append(
-                {
-                    "frame": matrix_ego_per_agent,
-                    "last_action": last_action,
-                    "last_action_success": last_action_success,
-                }
-            )
-        return states
