+++ cordialsync/learning/optimizer.py
@@ -1,3 +1,5 @@
+from __future__ import division
+
 import math
 import torch
 import torch.optim as optim
