+++ cordialsync/utilities/furnmove.py
@@ -4,7 +4,6 @@ import json
 import math
 import os
 import time
-import warnings
 from collections import defaultdict
 from typing import List, Dict, Any
 
@@ -18,178 +17,6 @@ from cordialsync.learning.agent import A3CAgent
 from cordialsync.learning.multiagent import MultiAgent
 from cordialsync.ai2thor.episodes import MultiAgentAI2ThorEpisode
 from cordialsync.utilities.visualization import get_agent_map_data, visualize_agent_path
-
-
-def step_results_to_data_to_save_for_furnmove_experiment(
-    episode: MultiAgentAI2ThorEpisode, step_results: List[List[Dict]]
-):
-    if len(step_results) == 0:
-        return {}
-
-    episode_info = episode.info()
-    nagents = len(step_results[0])
-    info = {
-        "ep_length": len(step_results) * nagents,
-        "final_distance": episode_info["navigation/final_distance"],
-        "reached_target": episode_info["navigation/reached_target"],
-        "spl_manhattan": episode_info["navigation/spl_manhattan"],
-        "initial_manhattan_steps": episode_info["navigation/initial_manhattan_steps"],
-    }
-    avail_actions = episode.available_actions
-
-    action_taken_matrix = np.zeros((len(avail_actions),) * nagents)
-    action_taken_success_matrix = np.zeros((len(avail_actions),) * nagents)
-
-    agent_action_info = [defaultdict(lambda: []) for _ in range(nagents)]
-    total_reward = 0
-    for srs in step_results:
-        action_inds = []
-        success_so_far = None
-
-        for agent_id, sr in enumerate(srs):
-            agent_info = agent_action_info[agent_id]
-            total_reward += sr["reward"]
-
-            action_ind = sr["action"]
-            action = avail_actions[sr["action"]]
-            action_success = sr["action_success"]
-            action_inds.append(action_ind)
-
-            agent_info["action_ind"].append(action_ind)
-            agent_info["action"].append(action)
-            agent_info["action_success"].append(action_success)
-
-            if "extra_before_info" in sr:
-                agent_info["dresser_visible"].append(
-                    sr["extra_before_info"]["dresser_visible"]
-                )
-                agent_info["tv_visible"].append(sr["extra_before_info"]["tv_visible"])
-
-            if success_so_far is None:
-                success_so_far = action_success
-            elif success_so_far != action_success:
-                success_so_far = math.nan
-
-        action_inds = tuple(action_inds)
-        if success_so_far is math.nan:
-            if all(
-                "Pass" != agent_action_info[i]["action"][-1] for i in range(nagents)
-            ):
-                warnings.warn("One agent was successful while the other was not!")
-        elif success_so_far:
-            action_taken_success_matrix[action_inds] += 1
-        action_taken_matrix[action_inds] += 1
-
-    info["step_results"] = step_results
-    info["reward"] = total_reward
-    info["agent_action_info"] = [dict(info) for info in agent_action_info]
-    info["action_taken_matrix"] = action_taken_matrix.tolist()
-    info["action_taken_success_matrix"] = action_taken_success_matrix.tolist()
-    return info
-
-
-class SaveFurnMoveMixin(object):
-    def simple_name(self):
-        raise NotImplementedError()
-
-    def create_episode_summary(
-        self,
-        agent: A3CAgent,
-        additional_metrics: Dict[str, float],
-        step_results: List[List[Dict]],
-    ) -> Any:
-        data_to_save = step_results_to_data_to_save_for_furnmove_experiment(
-            episode=agent.episode, step_results=step_results
-        )
-        data_to_save["episode_init_data"] = self.init_train_agent.episode_init_data
-
-        if len(agent.eval_results) != 0:
-            for i in range(agent.environment.num_agents):
-                for k in ["talk_probs", "reply_probs"]:
-                    data_to_save["a{}_{}".format(i, k)] = []
-                    for er in agent.eval_results:
-                        if k in er:
-                            data_to_save["a{}_{}".format(i, k)].append(
-                                er[k][i].cpu().detach().view(-1).numpy().tolist()
-                            )
-            if "randomization_probs" in agent.eval_results[0]:
-                data_to_save["randomization_probs"] = []
-                for er in agent.eval_results:
-                    data_to_save["randomization_probs"].append(
-                        torch.exp(er["randomization_log_probs"])
-                        .cpu()
-                        .detach()
-                        .view(-1)
-                        .numpy()
-                        .tolist()
-                    )
-
-        return {**data_to_save, **additional_metrics}
-
-    def data_save_dir(self) -> str:
-        return os.path.join(CONSTANTS.ABS_PATH_TO_DATA_DIR, "furnmove_evaluations__{}/{}").format(
-            self.episode_init_queue_file_name.split("__")[-1].replace(".json", ""),
-            self.simple_name(),
-        )
-
-    def save_episode_summary(self, data_to_save: Any):
-        self.total_saved += 1
-
-        scene = data_to_save["episode_init_data"]["scene"]
-        rep = data_to_save["episode_init_data"]["replication"]
-        print(
-            "Saving data corresponding to scene {} and replication {}. {}/{} completed.".format(
-                scene, rep, self.total_saved, self.total_to_save
-            ),
-            flush=True,
-        )
-
-        write_path = os.path.join(
-            self.data_save_dir(), "{}__{}.json".format(scene, rep)
-        )
-
-        with open(write_path, "w") as f:
-            json.dump(data_to_save, f)
-
-    def create_episode_init_queue(self, mp_module) -> torch.multiprocessing.Queue:
-        q = mp_module.Queue()
-        self.total_saved = 0
-
-        with open(
-            os.path.join(CONSTANTS.ABS_PATH_TO_DATA_DIR, "{}").format(
-                self.episode_init_queue_file_name
-            ),
-            "r",
-        ) as f:
-            scene_to_rep_to_ep_data = json.load(f)
-
-        keys_remaining = {
-            (scene, rep)
-            for scene in scene_to_rep_to_ep_data
-            for rep in scene_to_rep_to_ep_data[scene]
-        }
-
-        os.makedirs(self.data_save_dir(), exist_ok=True)
-
-        for p in glob.glob(os.path.join(self.data_save_dir(), "*.json")):
-            file_name = os.path.basename(p)
-            keys_remaining.remove(tuple(file_name.replace(".json", "").split("__")))
-
-        k = 0
-        for scene, rep in keys_remaining:
-            data = scene_to_rep_to_ep_data[scene][rep]
-            data["scene"] = scene
-            data["replication"] = rep
-            q.put(data)
-            k += 1
-
-        print("{} episodes added to queue".format(k))
-        self.total_to_save = k
-
-        return q
-
-    def stopping_criteria_reached(self):
-        return self.total_saved >= self.total_to_save
 
 
 def step_results_to_data_to_save_for_furnlift_experiment(
